% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_detect.R
\name{detection}
\alias{detection}
\title{ASE Detection}
\usage{
detection(data, clean_index, paras, rep, fdr = 0.05)
}
\arguments{
\item{data}{The raw data, need to be in the exact format with the sample data provided by this package, more specifically, a \code{data.frame} with its first 3 variables indicating gene ID, gene number and SNP number within a specific gene respectively. The following 2*Rep (Rep is the number of biological replicates) columns in the \code{data.frame} contain count data and the order must be consistent with the sample data. More details would be found in \code{help(mysample)}.}

\item{clean_index}{a vector contain index of genes without computational problems which would be returned by \code{par.est} function.}

\item{paras}{hyperparameters estimates returned by \code{par.est} function.}

\item{rep}{The number of biological replicates at each SNP.}

\item{fdr}{The desired False Discovery Rate (FDR), default value was set as 0.05.}
}
\value{
A \code{list} contains the following:
\describe{
\item{geneEffect}{A \code{data.frame} contains gene IDs and gene numbers of the genes with significant ASE and their corresponding posterior probabilities and estimated FDRs.}
\item{SNPEffect}{A \code{data.frame} contains gene IDs and gene numbers of the genes with significant ASE variation across SNPs and their corresponding posterior probabilities and estimated FDRs.}
\item{GSEffect}{A \code{data.frame} contains gene IDs and gene numbers of the genes exhibiting both ASE gene effect and ASE variation across SNPs and their corresponding posterior probabilities and estimated FDRs.}
\item{logPPs}{a \code{data.frame} contains detailed intermediate results, i.e., the posterior probabilities of the 4 models for all the genes.}
}
}
\description{
A function to perform hypothesis tests.
}
