% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDD.R
\name{GDD}
\alias{GDD}
\title{Re-structurize the input data}
\usage{
GDD(i, data, rep)
}
\arguments{
\item{i}{The gene number of each gene. Take the sample dataset for instance, the gene number of the first gene in this dataset is 28.}

\item{data}{The raw data, need to be in the exact format with the sample data provided by this package, more specifically, a \code{data.frame} with its first 3 variables indicating gene ID, gene number and SNP number within a specific gene respectively. The following 2*Rep (Rep is the number of biological replicates) columns in the \code{data.frame} contain count data and the order must be consistent with the sample data. More details would be found in \code{help(mysample)}.}

\item{rep}{The number of biological replicates at each SNP.}
}
\value{
A \code{data.frame} in the structure that would be needed for analysis by BLRM method.
}
\description{
A function to convert the raw data into a structure which contains necessary information for analysis, i.e., SNPs, Replicates, counts  from maternal allel (YI) and total counts (NI). This structure is friendly to GLMM fitting in glmer() function in "lme4" package. This function can be easily modified to apply to other similar real data structures.
}
