% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{mysample}
\alias{mysample}
\title{Sample data for Analysis}
\format{A \code{data.frame} with 2673 observations and 11 variables, which are:
\describe{
\item{EnsembleID}{the ensemble ID of each gene}
\item{GeneNum}{the number of genes}
\item{snp}{SNP of genes}
\item{sum1}{the total counts from maternal and paternal allels for the 1st biological replicates}
\item{t1}{the counts from maternal allel for the 1st biological replicates}
\item{sum2}{the total counts from maternal and paternal allels for the 2nd biological replicates}
\item{t2}{the counts from maternal allel for the 2nd biological replicates}
\item{sum3}{the total counts from maternal and paternal allels for the 3rd biological replicates}
\item{t3}{the counts from maternal allel for the 3rd biological replicates}
\item{sum4}{the total counts from maternal and paternal allels for the 4th biological replicates}
\item{t4}{the counts from maternal allel for the 4th biological replicates}
}}
\usage{
mysample
}
\description{
A simple sample dataset that is for illustration of the BLRM method.
}
\keyword{datasets}
