% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.est.R
\name{para.est}
\alias{para.est}
\title{Hyperparameters Estimation}
\usage{
para.est(data, rep)
}
\arguments{
\item{data}{The raw data, need to be in the exact format with the sample data provided by this package, more specifically, a \code{data.frame} with its first 3 variables indicating gene ID, gene number and SNP number within a specific gene respectively. The following 2*Rep (Rep is the number of biological replicates) columns in the \code{data.frame} contain count data and the order must be consistent with the sample data. More details would be found in \code{help(mysample)}.}

\item{rep}{The number of biological replicates at each SNP.}
}
\value{
A \code{list} contains the following:
\describe{
\item{para}{the estimation of hyperparameters.}
\item{index}{a vector contain index of genes without computational problems.}
\item{all}{a \code{data.frame} contains detailed intermediate results such as the p_values and estimated FDRs of likelihood ratio tests, the estimates of variance components etc.}
}
}
\description{
A function to estimate hyperparameters, i.e., the scale and location parameters of prior distributions (Gamma) of the variances of random replicates effect and random SNP effect, as well as the parameters of the prior distribution (Gaussian) of fixed gene effect.This function tried to fit Generalized Linear Mixed Model to the data of each gene by \code{glmer} function in \code{lme4} and then filter out the genes with computational problems. The hyperparameter estimation is based on the genes without computational problems.
}
